+++ src/main/java/cz/patyk/invoicesystem_be/repositories/InfluencingTicketRepository.java
@@ -3,5 +3,5 @@ package cz.patyk.invoicesystem_be.repositories;
 import cz.patyk.invoicesystem_be.entities.InfluencingTicket;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InfluencingTicketRepository extends JpaRepository<InfluencingTicket, Integer> {
+public interface InfluencingTicketRepository extends JpaRepository<InfluencingTicket, Long> {
 }
