+++ src/main/java/cz/patyk/invoicesystem_be/repositories/InvoiceItemRepository.java
@@ -3,5 +3,5 @@ package cz.patyk.invoicesystem_be.repositories;
 import cz.patyk.invoicesystem_be.entities.InvoiceItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InvoiceItemRepository extends JpaRepository<InvoiceItem, Integer> {
+public interface InvoiceItemRepository extends JpaRepository<InvoiceItem, Long> {
 }
