+++ src/main/java/cz/patyk/invoicesystem_be/repositories/PaymentTypeRepository.java
@@ -3,5 +3,5 @@ package cz.patyk.invoicesystem_be.repositories;
 import cz.patyk.invoicesystem_be.entities.PaymentType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PaymentTypeRepository extends JpaRepository<PaymentType, Integer> {
+public interface PaymentTypeRepository extends JpaRepository<PaymentType, Long> {
 }
