+++ src/main/java/cz/patyk/invoicesystem_be/repositories/TicketTypeRepository.java
@@ -3,5 +3,5 @@ package cz.patyk.invoicesystem_be.repositories;
 import cz.patyk.invoicesystem_be.entities.TicketType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TicketTypeRepository extends JpaRepository<TicketType, Integer> {
+public interface TicketTypeRepository extends JpaRepository<TicketType, Long> {
 }
